pyproject.toml
[project]
name = "recipes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.10.6",
    "pytest>=8.3.4",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
